Remove support for end-of-life Pythons
Python < 3.5 is end of life. They are no longer receiving bug fixes,
including for security issues. Here is a list of end of life Pythons and
their dates:

https://devguide.python.org/devcycle/#end-of-life-branches

By removing support for these, the project is in a better position to
adopt newer Python features. For example, type annotations.

This also reduces testing and support resources.